// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/order-creation-request.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderDemand } from './order-demand';


export interface OrderCreationRequest { 
    /**
     * Name of the template to be used for the order request.
     */
    templateName: string;
    isDraft?: boolean;
    demands?: Array<OrderDemand>;
}