// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/order-demand.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderDemandType } from './order-demand-type';
import { DemandLocation } from './demand-location';


export interface OrderDemand { 
    demandLocation: DemandLocation;
    /**
     * Demand date and time.
     */
    demandDateTime: string;
    demandTypes: Array<OrderDemandType>;
}