// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/order-demand-type.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NHM } from './nhm';
import { DemandWagonType } from './demand-wagon-type';
import { TransitRailwayUndertaking } from './transit-railway-undertaking';
import { CommercialLocation } from './commercial-location';


export interface OrderDemandType { 
    /**
     * Maximum amount of wagons that can be ordered of this demand type.
     */
    maxOrderAmount?: number;
    assessment?: boolean;
    shipperAuthorization?: boolean;
    commentCustomer?: string;
    referenceNumber?: string;
    demandWagonType: DemandWagonType;
    nhm?: NHM;
    loadRunLocation?: CommercialLocation;
    transitRailwayUndertaking?: TransitRailwayUndertaking;
    /**
     * Number of wagons to be ordered.
     */
    numberOfWagons: number;
    /**
     * Reference number for the orders to be created chosen by the customer.
     */
    customerReference?: string;
    /**
     * Comment addressed to the customer service.
     */
    commentToCustomerService?: string;
}